Fixed things

+++ differential_privacy/dataset.py
@@ -9,11 +9,12 @@ class Dataset:
     def get(self):
         return self.x, self.y
     
+    # Cambiarlo
     def get_split(index: int, split_count: int):
         pass
 
     #Propongo hacer un método get_generalisation_fragment que te de el split de cada dataset
-    def get_generalisation_fragment(fraction: float):
+    def get_generalisation_fragment(fraction: int):
         fragment_size = len(self.x) // fraction
         return Dataset(x[:fragment_size], y[:fragment_size])
